# docs/swagger.yaml
basePath: /api
definitions:
...
  internal_handlers.CreateProcedureRequest:
    properties:
      anam:
        type: string
      coment:
        type: string
      dani:
        type: string
      date:
        type: string
      date2:
        type: string
      date3:
        type: string
      deh:
        type: string
      diagn:
        type: string
      koment:
        type: string
      nout:
        type: string
      owner:
        type: string
      ownern:
        type: string
      phone:
        type: string
      pname:
        type: string
      price:
        type: string
      ser:
        type: string
      tp:
        minimum: 1
        type: integer
      tpname:
        type: string
      uuid:
        type: string
      vac:
        type: string
      vac1:
        type: string
      vac2:
        type: string
      vac3:
        type: string
      vac4:
        type: string
      vac5:
        type: string
      vac6:
        type: string
      vac7:
        type: string
      vac8:
        type: string
      vac9:
        type: string
      vacn:
        type: string
    required:
    - tp
    - uuid
    type: object
...
  internal_handlers.EctoOptionsResponse:
    properties:
      collars:
        items:
          $ref: '#/definitions/internal_handlers.SelectOption'
        type: array
      drops:
        items:
          $ref: '#/definitions/internal_handlers.SelectOption'
        type: array
      sprays:
        items:
          $ref: '#/definitions/internal_handlers.SelectOption'
        type: array
      tablets:
        items:
          $ref: '#/definitions/internal_handlers.SelectOption'
        type: array
    required:
    - collars
    - drops
    - sprays
    - tablets
    type: object
...
  internal_handlers.ErrorResponse:
    properties:
      error:
        type: string
    required:
    - error
    type: object
...
  internal_handlers.MessageResponse:
    properties:
      message:
        type: string
    required:
    - message
    type: object
...
  internal_handlers.ProcedureTypeItem:
    properties:
      name:
        type: string
      tp:
        type: integer
    required:
    - name
    - tp
    type: object
...
  internal_handlers.SelectOption:
    properties:
      label:
        type: string
      value:
        type: string
    required:
    - label
    - value
    type: object
...
  internal_handlers.VaccineOptionsResponse:
    properties:
      brands:
        items:
          $ref: '#/definitions/internal_handlers.SelectOption'
        type: array
      vaccines:
        items:
          $ref: '#/definitions/internal_handlers.SelectOption'
        type: array
    required:
    - brands
    - vaccines
    type: object
...
  vetapp-backend_internal_models.Procedure:
    properties:
      address:
        description: |-
          Denormalized identity / housekeeping columns the legacy table also
          stores. Mostly redundant with Pet but the PHP forms set them on
          every save and some queries depend on them.
        type: string
      anam:
        description: Anamnesis
        type: string
      coment:
        description: Comment visible to owner
        type: string
      company:
        description: payment method label e.g. "ბარათი", "ნაღდი"
        type: string
      dani:
        description: Prescription
        type: string
      date:
        description: Procedure date
        type: string
      date2:
        description: Next due date
        type: string
      date3:
        description: SMS reminder trigger date (= date2 - 1 month)
        type: string
      deh:
        description: Dehel drug-from-list (tp=12) OR Canine Babesia result (tp=2)
        type: string
      diagn:
        description: Diagnosis
        type: string
      id:
        type: integer
      koment:
        description: Vet-internal notes
        type: string
      name:
        description: edit-log message column (NOT the pet name — that's PName)
        type: string
      nout:
        description: Treatment
        type: string
      owner:
        description: Owner personal ID
        type: string
      ownern:
        description: Owner name (denormalized)
        type: string
      phone:
        description: 'Payment status: "0"=unpaid, "1"=paid'
        type: string
      pn:
        description: owner personal-id copy
        type: string
      pname:
        description: Pet name (denormalized)
        type: string
      price:
        description: Price in GEL
        type: string
      sax:
        description: pet sex denormalized
        type: string
      ser:
        description: Vaccine serial / batch number
        type: string
      sk:
        description: Clinic zip
        type: string
      test1:
        description: |-
          Extra test-result columns specific to dog tests (tp=2). PHP form
          at vet/addtest.php uses these for the Caniv 4DX panel + CDV/CAV
          test slots. Other tps leave them empty.
        type: string
      test2:
        type: string
      test3:
        type: string
      test4:
        type: string
      test5:
        type: string
      test6:
        type: string
      test7:
        type: string
      test8:
        type: string
      tp:
        description: Procedure type code
        type: integer
      tpname:
        description: Procedure type display name
        type: string
      uuid:
        description: Pet ID (as string)
        type: string
      vac:
        description: Vaccine/procedure name OR tp=11 drops-custom OR tp=12 custom
          drug
        type: string
      vac1:
        description: |-
          Polymorphic content slots. For tp=11 (ecto): drops/pills/collar/
          spray with custom-text in Vac/Vac2/Vac4/Vac6 and dropdown choice
          in Vac1/Vac3/Vac5/Vac7. For tp=2/22/222: test panel results. For
          tp=10x/20x (generic): Vac1=anamnesis, Vac2=diagnosis, Vac3=treatment.
          Vac8/Vac9 are unused in production data (empty in all rows we've
          inspected) but kept here for legacy compatibility.
        type: string
      vac2:
        type: string
      vac3:
        type: string
      vac4:
        type: string
      vac5:
        type: string
      vac6:
        type: string
      vac7:
        type: string
      vac8:
        type: string
      vac9:
        type: string
      vacn:
        description: Brand OR tp=2 Leishmania result
        type: string
      vetname:
        description: Vet member ID (numeric string) — empty/0 means owner-added
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Veterinary clinic management API
  title: VetApp API
  version: "1.0"
paths:
...
  /procedures:
    get:
      parameters:
      - description: Clinic code
        in: query
        name: clinic
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      - description: Vet member ID
        in: query
        name: vet_id
        type: string
      - description: Pet ID
        in: query
        name: pet_id
        type: string
      - description: Procedure type code
        in: query
        name: tp
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/vetapp-backend_internal_models.Procedure'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List procedures
      tags:
      - procedures
    post:
      consumes:
      - application/json
      parameters:
      - description: Procedure data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateProcedureRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/vetapp-backend_internal_models.Procedure'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create procedure
      tags:
      - procedures
  /procedures/{id}:
    delete:
      parameters:
      - description: Procedure ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete procedure
      tags:
      - procedures
    get:
      parameters:
      - description: Procedure ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vetapp-backend_internal_models.Procedure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get procedure by ID
      tags:
      - procedures
    put:
      consumes:
      - application/json
      parameters:
      - description: Procedure ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vetapp-backend_internal_models.Procedure'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update procedure
      tags:
      - procedures
  /procedures/dehel-options:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.SelectOption'
            type: array
      security:
      - BearerAuth: []
      summary: List dehelminization options
      tags:
      - procedures
  /procedures/ecto-options:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.EctoOptionsResponse'
      security:
      - BearerAuth: []
      summary: List ectoparasite options
      tags:
      - procedures
  /procedures/test-options:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.SelectOption'
            type: array
      security:
      - BearerAuth: []
      summary: List test options
      tags:
      - procedures
  /procedures/types:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.ProcedureTypeItem'
            type: array
      security:
      - BearerAuth: []
      summary: List procedure types
      tags:
      - procedures
  /procedures/vaccine-options:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.VaccineOptionsResponse'
      security:
      - BearerAuth: []
      summary: List vaccine options
      tags:
      - procedures
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"